# conectatech-backend/docs/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ConectaTech API",
    "description": "API que conecta estudantes de diversos cursos da UNP a vagas de estagio na regiao de Mossoro/RN",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Listar vagas",
        "description": "Lista todas as vagas com filtros opcionais por texto, localizacao e fonte",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "description": "Texto para buscar em titulo, empresa e descricao",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "localizacao",
            "in": "query",
            "description": "Filtrar por localizacao (ex: Mossoro)",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "fonte",
            "in": "query",
            "description": "Filtrar por fonte: JOOBLE, JSOUP ou ADMIN",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de vagas retornada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VagaDTO" }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "tags": ["Vagas"],
        "summary": "Buscar vaga por ID",
        "description": "Retorna os detalhes de uma vaga especifica",
        "operationId": "getJobById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da vaga",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vaga encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VagaDTO" }
              }
            }
          },
          "404": {
            "description": "Vaga nao encontrada"
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Listar usuarios",
        "description": "Retorna todos os usuarios cadastrados",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UsuarioDTO" }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado - requer role ADMIN"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Buscar usuario por ID",
        "description": "Retorna os dados de um usuario especifico",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuario",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioDTO" }
              }
            }
          },
          "404": {
            "description": "Usuario nao encontrado"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "tags": ["Admin"],
        "summary": "Atualizar usuario",
        "description": "Atualiza os dados de um usuario",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuario",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UsuarioDTO" }
              }
            }
          },
          "404": {
            "description": "Usuario nao encontrado"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/jobs": {
      "get": {
        "tags": ["Admin"],
        "summary": "Listar todas as vagas (admin)",
        "description": "Retorna todas as vagas cadastradas no sistema",
        "operationId": "listAllJobs",
        "responses": {
          "200": {
            "description": "Lista de vagas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VagaDTO" }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Criar vaga manualmente",
        "description": "Cadastra uma nova vaga com fonte ADMIN",
        "operationId": "createJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VagaDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vaga criada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VagaDTO" }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/jobs/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Atualizar vaga",
        "description": "Atualiza os dados de uma vaga existente",
        "operationId": "updateJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da vaga",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VagaDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vaga atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VagaDTO" }
              }
            }
          },
          "404": {
            "description": "Vaga nao encontrada"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Deletar vaga",
        "description": "Remove permanentemente uma vaga do sistema",
        "operationId": "deleteJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da vaga",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Vaga removida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Vaga removida com sucesso" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vaga nao encontrada"
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/import-scraper": {
      "post": {
        "tags": ["Admin"],
        "summary": "Executar scraping",
        "description": "Executa o scraper manualmente para importar vagas de sites externos",
        "operationId": "importScraper",
        "responses": {
          "200": {
            "description": "Scraping executado com quantidade de vagas importadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Scraping executado" },
                    "vagasImportadas": { "type": "integer", "example": 15 }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/api/admin/import-jooble": {
      "post": {
        "tags": ["Admin"],
        "summary": "Importar vagas do Jooble",
        "description": "Busca e importa vagas da API Jooble para o banco de dados",
        "operationId": "importJooble",
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "description": "Palavras-chave para busca",
            "required": false,
            "schema": { "type": "string", "default": "estagio tecnologia" }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Localizacao da busca",
            "required": false,
            "schema": { "type": "string", "default": "Mossoro, RN" }
          }
        ],
        "responses": {
          "200": {
            "description": "Importacao concluida com quantidade de vagas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Importacao Jooble concluida" },
                    "vagasImportadas": { "type": "integer", "example": 10 }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  },
  "components": {
    "schemas": {
...
      "UsuarioDTO": {
        "required": ["nome"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do usuario", "example": 1 },
          "nome": { "type": "string", "description": "Nome completo", "example": "Joao Silva" },
          "email": { "type": "string", "description": "Email", "example": "joao.silva@email.com" },
          "curso": { "type": "string", "description": "Curso", "example": "Ciencia da Computacao" },
          "periodo": { "type": "string", "description": "Periodo", "example": "6" },
          "role": { "type": "string", "description": "Role do usuario", "example": "STUDENT" },
          "criadoEm": { "type": "string", "description": "Data de criacao", "example": "30/03/2026 14:30" }
        },
        "description": "Dados do usuario"
      },
      "VagaDTO": {
        "required": ["titulo", "empresa"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da vaga", "example": 1 },
          "titulo": { "type": "string", "description": "Titulo da vaga", "example": "Estagiario de Desenvolvimento Web" },
          "empresa": { "type": "string", "description": "Nome da empresa", "example": "Tech Solutions LTDA" },
          "descricao": { "type": "string", "description": "Descricao da vaga", "example": "Vaga para desenvolvimento web com React e Spring Boot" },
          "localizacao": { "type": "string", "description": "Localizacao da vaga", "example": "Mossoro, RN" },
          "url": { "type": "string", "description": "URL da vaga original", "example": "https://exemplo.com/vaga/123" },
          "fonte": { "type": "string", "description": "Fonte da vaga", "example": "JOOBLE" },
          "dataPublicacao": { "type": "string", "description": "Data de publicacao", "example": "30/03/2026 10:00" }
        },
        "description": "Dados de uma vaga de estagio"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Token JWT obtained via POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}